requirements.txt
alembic==1.13.1
asyncpg==0.29.0
black==24.1.1
fastapi==0.109.2
flake8==7.0.0
google-generativeai==0.3.2
isort==5.13.2
mypy==1.8.0

# Development dependencies
pre-commit==3.6.0
psycopg2-binary==2.9.9
pydantic==2.6.1
pydantic-settings==2.1.0
python-dotenv==1.0.1
python-jose==3.3.0
sqlalchemy==2.0.27
types-python-jose==3.3.4.20240106
uuid6==2024.1.12
uvicorn==0.27.1
